# route-service-fleetsync/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FleetSync Route Service API",
        "description": "API for managing routes in FleetSync application",
        "contact": {
            "email": "support@fleetsync.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/routes": {
            "get": {
                "tags": [
                    "Routes"
                ],
                "summary": "List all routes",
                "operationId": "25d3d0d0f5d12b021dc6272e30293e95",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Route"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch routes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Routes"
                ],
                "summary": "Create a new route",
                "operationId": "b33c4c1179bc9292122b3d06edfeaff1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "driver_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vehicle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_location": {
                                        "type": "string",
                                        "example": "Jakarta"
                                    },
                                    "end_location": {
                                        "type": "string",
                                        "example": "Bandung"
                                    },
                                    "distance_km": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.5
                                    },
                                    "estimated_time_minutes": {
                                        "type": "integer",
                                        "example": 180
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "planned"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Route created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Route"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Failed to create route"
                    }
                }
            }
        },
        "/api/routes/{id}": {
            "get": {
                "tags": [
                    "Routes"
                ],
                "summary": "Get route by ID",
                "operationId": "e2a8d29a88e0b4b65e46b742ac68f79e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Route"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Route not found"
                    },
                    "500": {
                        "description": "Failed to fetch route"
                    }
                }
            },
            "put": {
                "tags": [
                    "Routes"
                ],
                "summary": "Update a route",
                "operationId": "deeadf381bc7b3d6471d4c17c0cfe8ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "driver_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vehicle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_location": {
                                        "type": "string",
                                        "example": "Jakarta"
                                    },
                                    "end_location": {
                                        "type": "string",
                                        "example": "Bandung"
                                    },
                                    "distance_km": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.5
                                    },
                                    "estimated_time_minutes": {
                                        "type": "integer",
                                        "example": 180
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "in_progress"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Route updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Route"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Route not found"
                    },
                    "500": {
                        "description": "Failed to update route"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Routes"
                ],
                "summary": "Delete a route",
                "operationId": "90c67fe2765db8d58d5efc5a24316169",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Route deleted"
                    },
                    "404": {
                        "description": "Route not found"
                    },
                    "500": {
                        "description": "Failed to delete route"
                    }
                }
            }
        },
        "/api/routes/{id}/status": {
            "post": {
                "tags": [
                    "Routes"
                ],
                "summary": "Update route status",
                "operationId": "48f54911435e039ce27355573943077c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Route"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status"
                    },
                    "404": {
                        "description": "Route not found"
                    },
                    "500": {
                        "description": "Failed to update status"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Route": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "driver_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "vehicle_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "start_location": {
                        "type": "string",
                        "example": "Jakarta"
                    },
                    "end_location": {
                        "type": "string",
                        "example": "Bandung"
                    },
                    "distance_km": {
                        "type": "number",
                        "format": "float",
                        "example": 150.5
                    },
                    "estimated_time_minutes": {
                        "type": "integer",
                        "example": 180
                    },
                    "status": {
                        "type": "string",
                        "example": "planned"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Routes",
            "description": "Routes"
        }
    ]
}